ejemplos/ejemplo-agente-con-streaming/pyproject.toml
[project]
name = "ejemplo-agente-con-streaming"
version = "0.1.0"
description = "Ejemplo de agente LangGraph con streaming de tokens"
dependencies = [
    "langgraph",
    "langchain-openai",
    "langchain-core",
    "python-dotenv",
    "langgraph-cli[inmem]"
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
